RavenDB-21366 - move and simplify test
